requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-multipart==0.0.6
